requirements.txt
# Core dependencies for Overcooked MARL Project
# Note: Install PyTorch separately from pytorch.org based on your system

# Overcooked environment
overcooked-ai==1.1.0

# Deep learning (install from pytorch.org instead)
# torch>=2.0.0
# torchvision>=0.15.0

# Data handling and visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
tensorboard>=2.13.0

# Optional: For better plotting
plotly>=5.14.0

# Development
jupyter>=1.0.0
ipython>=8.12.0
